+++ src/app/(providers)/(root)/challenges/[id]/detail/page.tsx
@@ -1,13 +1,10 @@
 'use client';
 import { useGetUser } from '@/hooks/auth/useUsers';
-import { useChallengeUpdate, useGetChallengeDetail } from '@/hooks/challenge/useChallenge';
-import { useImageUpload } from '@/hooks/image/useImage';
+import { useGetChallengeDetail } from '@/hooks/challenge/useChallenge';
 import { useGetReviews } from '@/hooks/review/useReview';
-import { Tables } from '@/types/supabase';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
-import { FormEvent, useRef, useState } from 'react';
-import { FormFields } from '../../register/_components/ChallengeRegisterForm/ChallengeRegisterForm';
+import { useRef, useState } from 'react';
 
 const ChallengeDetailPage = ({ params }: { params: { id: string } }) => {
   const inputRef = useRef<HTMLInputElement>(null);
@@ -16,96 +13,12 @@ const ChallengeDetailPage = ({ params }: { params: { id: string } }) => {
   const { data: user } = useGetUser();
   const { data: challenge } = useGetChallengeDetail(id);
   const { data: reviews } = useGetReviews(id);
-  const { mutate: imageUpload } = useImageUpload();
-  const { mutate: challengeUpdate } = useChallengeUpdate();
   const router = useRouter();
 
   //TODO: 로딩
   if (!challenge) {
     return <div>없따!</div>;
   }
-
-  const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    const file = inputRef?.current?.files?.[0] || null;
-
-    if (!file && !challenge.imageURL) {
-      console.error('Challenge Register Image Error : 사진을 올려주세요.');
-      return;
-    }
-
-    const formData = new FormData(e.currentTarget);
-    const fields: (keyof FormFields)[] = ['title', 'content', 'startDate', 'endDate', 'category'];
-    const formFields: Partial<FormFields> = {};
-
-    for (const field of fields) {
-      const value = formData.get(field);
-      if (typeof value !== 'string' || value.trim() === '') {
-        console.error(`Challenge Register ${field} Error : ${field}을(를) 입력 해주세요.`);
-        return;
-      }
-      formFields[field] = value.trim();
-    }
-
-    const { title, content, startDate, endDate, category } = formFields as FormFields;
-
-    if (file) {
-      const form = new FormData();
-      form.append('file', file);
-      imageUpload(
-        { storage: 'challengeRegister', form },
-        {
-          onSuccess: (response) => {
-            const updateData: Omit<Tables<'challenges'>, 'id'> = {
-              title,
-              content,
-              startDate,
-              endDate,
-              isProgress: challenge.isProgress,
-              createdBy: challenge.createdBy,
-              imageURL: response.imageURL,
-              verify: null,
-              tags: null,
-              rating: 0,
-              category,
-            };
-            challengeUpdate(
-              { updateData, cid: challenge.id },
-              {
-                onSuccess: () => {
-                  alert('수정이 완료되었습니다.');
-                  router.replace(`/challenges/discover`);
-                },
-              },
-            );
-          },
-        },
-      );
-    } else if (challenge.imageURL) {
-      const updateData: Omit<Tables<'challenges'>, 'id'> = {
-        title,
-        content,
-        startDate,
-        endDate,
-        isProgress: challenge.isProgress,
-        createdBy: challenge.createdBy,
-        imageURL: challenge.imageURL,
-        verify: null,
-        tags: null,
-        rating: 0,
-        category,
-      };
-      challengeUpdate(
-        { updateData, cid: challenge.id },
-        {
-          onSuccess: () => {
-            alert('수정이 완료되었습니다.');
-            router.replace(`/challenges/discover`);
-          },
-        },
-      );
-    }
-  };
 
   return (
     <div className="h-screen">
